moved molecule/slides to template/app-player/player/slides
