+++ packages/@coorpacademy-components/src/index.js
@@ -84,8 +84,6 @@ export SetupSection from './molecule/setup-section';
 export SetupSections from './molecule/setup-sections';
 export SetupSlide from './molecule/setup-slide';
 export SetupSlider from './molecule/setup-slider';
-export SlidesFooter from './molecule/slides/slides-footer';
-export SlidesPlayer from './molecule/slides/slides-player';
 export SsmenuList from './molecule/ssmenu-list';
 export SubscriptionFreemium from './molecule/subscription-freemium';
 export SubscriptionPremium from './molecule/subscription-premium';
@@ -118,6 +116,8 @@ export Slider from './organism/slider';
 export Activity from './template/activity';
 export Loading from './template/app-player/loading';
 export PlayerHeader from './template/app-player/player/player-header';
+export SlidesFooter from './template/app-player/player/slides/slides-footer';
+export SlidesPlayer from './template/app-player/player/slides/slides-player';
 export Player from './template/app-player/player';
 export PopinCorrection from './template/app-player/popin-correction';
 export PopinEnd from './template/app-player/popin-end';
