+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-footer/test/fixtures/default.js
