+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-footer/test/fixtures/highlighted.js
