+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-footer/test/slides-footer.test.js
