+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/style.css
@@ -1,6 +1,6 @@
-@value breakpoints: "../../../variables/breakpoints.css";
+@value breakpoints: "../../../../../variables/breakpoints.css";
 @value mobile from breakpoints;
-@value colors: "../../../variables/colors.css";
+@value colors: "../../../../../variables/colors.css";
 @value medium from colors;
 @value dark from colors;
 @value light from colors;
