+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/clue.js
@@ -1,6 +1,6 @@
-import Cta from '../../../../../atom/cta/test/fixtures/primary';
-import Footer from '../../../slides-footer/test/fixtures/only-clue';
-import Clue from '../../../../../atom/clue/test/fixtures/default';
+import Cta from '../../../../../../../atom/cta/test/fixtures/primary';
+import Footer from '../../../slides-footer/test/fixtures/clue-selected';
+import Clue from '../../../../../../../atom/clue/test/fixtures/default';
 
 const {props} = Cta;
 const footerProps = Footer.props;
@@ -13,14 +13,12 @@ export default {
       current: 2,
       total: 6
     },
-    showNewMedia: false,
     question:
       'Amongst these businesses, which have suffered setbacks for not knowing how to putting users first?',
     cta: {
       ...props,
       submitValue: 'Back to Question'
     },
-    help: 'En regardant cet indice, vous perdez 1 étoile',
     ...clueProps,
     ...footerProps
   }
