+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/error.js
