+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/loading.js
@@ -1,4 +1,4 @@
-import Loader from '../../../../../atom/loader/test/fixtures/default';
+import Loader from '../../../../../../../atom/loader/test/fixtures/default';
 
 const loaderProps = Loader.props;
