+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/no-total.js
