+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/qcm-drag.js
@@ -1,4 +1,4 @@
-import Answer from '../../../../answer/test/fixtures/template';
+import Answer from '../../../../../../../molecule/answer/test/fixtures/qcm-drag';
 import Default from './default';
 
 const answerType = Answer.props;
@@ -7,6 +7,7 @@ export default {
   props: {
     typeClue: 'answer',
     ...Default.props,
+    help: null,
     answerType
   }
 };
