+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/qcm-template.js
@@ -1,4 +1,4 @@
-import Answer from '../../../../answer/test/fixtures/default';
+import Answer from '../../../../../../../molecule/answer/test/fixtures/template';
 import Default from './default';
 
 const answerType = Answer.props;
