+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/qcm.js
@@ -1,4 +1,4 @@
-import Answer from '../../../../answer/test/fixtures/qcm';
+import Answer from '../../../../../../../molecule/answer/test/fixtures/qcm';
 import Default from './default';
 
 const answerType = Answer.props;
