+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/range.js
@@ -1,4 +1,4 @@
-import Answer from '../../../../answer/test/fixtures/qcm-drag';
+import Answer from '../../../../../../../molecule/answer/test/fixtures/range';
 import Default from './default';
 
 const answerType = Answer.props;
@@ -7,7 +7,6 @@ export default {
   props: {
     typeClue: 'answer',
     ...Default.props,
-    help: null,
     answerType
   }
 };
