+++ packages/@coorpacademy-components/src/template/app-player/player/slides/slides-player/test/fixtures/with-min-height.js
