+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/clue.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import Clue from '../../../../../molecule/slides/slides-player/test/fixtures/clue';
+import Clue from '../../slides/slides-player/test/fixtures/clue';
 
 const headerProps = SlidesHeader.props;
 const playerProps = Clue.props;
