+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/context-with-image.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import Context from '../../../../../molecule/slides/slides-player/test/fixtures/context-with-image';
+import Context from '../../slides/slides-player/test/fixtures/context-with-image';
 
 const headerProps = SlidesHeader.props;
 const playerProps = Context.props;
