+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/context-with-video.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import Context from '../../../../../molecule/slides/slides-player/test/fixtures/context-with-video';
+import Context from '../../slides/slides-player/test/fixtures/context-with-video';
 
 const headerProps = SlidesHeader.props;
 const playerProps = Context.props;
