+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/drop-down.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import DropDown from '../../../../../molecule/slides/slides-player/test/fixtures/drop-down';
+import DropDown from '../../slides/slides-player/test/fixtures/drop-down';
 
 const headerProps = SlidesHeader.props;
 const playerProps = DropDown.props;
