+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/free-text.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import FreeText from '../../../../../molecule/slides/slides-player/test/fixtures/free-text';
+import FreeText from '../../slides/slides-player/test/fixtures/free-text';
 
 const headerProps = SlidesHeader.props;
 const playerProps = FreeText.props;
