+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/media.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import Media from '../../../../../molecule/slides/slides-player/test/fixtures/media';
+import Media from '../../slides/slides-player/test/fixtures/media';
 
 const headerProps = SlidesHeader.props;
 const playerProps = Media.props;
