+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/qcm-drag.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import QcmDrag from '../../../../../molecule/slides/slides-player/test/fixtures/qcm-drag';
+import QcmDrag from '../../slides/slides-player/test/fixtures/qcm-drag';
 
 const headerProps = SlidesHeader.props;
 const playerProps = QcmDrag.props;
