+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/qcm-graphic.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import QcmGraphic from '../../../../../molecule/slides/slides-player/test/fixtures/qcm-graphic';
+import QcmGraphic from '../../slides/slides-player/test/fixtures/qcm-graphic';
 
 const headerProps = SlidesHeader.props;
 const playerProps = QcmGraphic.props;
