+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/qcm-short.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import Qcm from '../../../../../molecule/slides/slides-player/test/fixtures/qcm-short';
+import Qcm from '../../slides/slides-player/test/fixtures/qcm-short';
 
 const headerProps = SlidesHeader.props;
 const playerProps = Qcm.props;
