+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/range.js
@@ -1,5 +1,5 @@
 import SlidesHeader from '../../player-header/test/fixtures/learner';
-import QuestionRange from '../../../../../molecule/slides/slides-player/test/fixtures/range';
+import QuestionRange from '../../slides/slides-player/test/fixtures/range';
 
 const headerProps = SlidesHeader.props;
 const playerProps = QuestionRange.props;
